+++ R/QFeatures-class.R
@@ -798,6 +798,7 @@ setReplaceMethod("names", c("QFeatures", "character"),
 ##'     (default `1L`)
 ##'
 ##' @importFrom MultiAssayExperiment longFormat
+##' @importFrom reshape2 melt
 ##'
 ##' @export
 longFormat <- function(object,
@@ -821,7 +822,7 @@ longFormat <- function(object,
         rds <- do.call(rbind, rds)
         ## Merge the rowData to the long table
         cbind(longDataFrame,
-              rds[longDataFrame$rowname, , drop = FALSE])
+              rds[as.character(longDataFrame$rowname), , drop = FALSE])
     } else {
         ## If rowvars is null, return the MAE longFormat output
         MultiAssayExperiment::longFormat(object, colvars, index)